pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Layash Humagai <layashhumagai@lftechnology.com>"]
readme = "README.md"

[tool.poetry.scripts]
my-script = "my_module:my-app"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
jinja2 = "^3.1.3"
pendulum = "^1.4"
mysql-connector-python = "^8.3.0"
werkzeug = "^3.0.1"
pyjwt = "^2.8.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
